Convert JQ Execute to using channels.
Given that JQ operates on streams already (you send one input, it
returns 0 or more outputs or errors, then you send it the next input and
it repeats until you close the input stream) channels seem like a nice
way of representing this in Go.

Structuring the code this way also provides clear boundaries of
responsibility and makes it easier for when we want to send multiple
values again and again once we have the REPL started.