Create a simpler interface for executing JQ programs
The channel interface was "nice" for my use in jqrepl (where I'm running
with the same input again and again but a different program each time)
but it doesn't make it easy to work out what input produced any given
output.

To make that possible add an "immediate mode" that simple returns takes
the input and returns the values directly.